requirements.txt
Flask>=2.3.0
Flask-Cors>=3.0.10
Flask-SQLAlchemy>=3.0.5
bcrypt>=4.0.1
PyJWT>=2.8.0
tensorflow>=2.14.0
Pillow>=10.0.0
numpy>=1.26.0
